Pull Request for #1: Design package
Closes #1

I've created a basic skeleton code using docstrings of the main functions that addresses the three outputs as described in #1. I've tried my best to adhere to the naming conventions in the paper, so hopefully this helps to see how the functions are supposed to interact with each other. Please let me know if anything is unclear or if you have any suggestions. Thanks!!